Implement cassandra table for syntactical checks

+++ common/src/main/scala/hmda/model/filing/ts/TransmittalLar.scala
@@ -1,6 +1,6 @@
 package hmda.model.filing.ts
 
-case class TransmittalLar(ts: TransmittalSheet, larsCount: Int = 0, larsDistinctCount: Int = 0, distinctUliCount: Int = 0)
-//S304 => ts.totalLines == larsCount
-//S305 => larsCount == larsDistinctCount
-//Q600 => larsCount == distinctUliCount
+case class TransmittalLar(ts: TransmittalSheet,
+                          larsCount: Int = 0,
+                          larsDistinctCount: Long = 0L,
+                          distinctUliCount: Long = 0L)
