+++ hmda/src/main/scala/hmda/validation/rules/lar/syntactical/S305.scala
@@ -10,7 +10,7 @@ object S305 extends EditCheck[TransmittalLar] {
   override def name: String = "S305"
   override def apply(tsLar: TransmittalLar): ValidationResult = {
 
-    tsLar.larsCount is equalTo(tsLar.larsDistinctCount)
+    tsLar.larsCount is equalTo(tsLar.larsDistinctCount.toInt)
 
   }
